Move stop hook, autostart to separate file.
In general, it'd be good to move separate units of functionality to
separate python files. That will make it easier to manage working on
those features across our charms.
